requirements.txt
Flask==2.3.2
transformers==4.30.2
sounddevice==0.4.6
wavio==0.0.7
google-cloud-speech==2.22.0
openai==0.27.6
